Babel module resolver added to simplify module paths in unit tests. Unit test added for presence.
